pyproject.toml
[project]
name = "medical-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.133.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "pypdf>=6.7.3",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "twilio>=9.10.2",
    "uvicorn>=0.41.0",
]
